Rename to p2p-util.js
